# src/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Hobit Admin API
  version: 1.0.0
  description: Hobit Admin API 설명
paths:
...
  /api/faqs/update/category/conflict:
    post:
      summary: FAQ 수정시 카테고리 충돌 여부 확인
      description: 수정하려는 FAQ의 카테고리 충돌 여부를 확인하는 API입니다.
      operationId: updateCheckFAQCategoryConflict
      tags:
        - faqs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                faq_id:
                  type: integer
                  description: FAQ ID
                  example: 1
                maincategory_ko:
                  type: string
                  description: 주요 카테고리 (한국어)
                  example: "주요 카테고리"
                maincategory_en:
                  type: string
                  description: 주요 카테고리 (영어)
                  example: "main category"
                subcategory_ko:
                  type: string
                  description: 하위 카테고리 (한국어)
                  example: "하위 카테고리"
                subcategory_en:
                  type: string
                  description: 하위 카테고리 (영어)
                  example: "sub category"
      responses:
        '200':
          description: 카테고리 확인 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: "Categories checked successfully"
                  data:
                    type: object
                    properties:
                      isConflict:
                        type: boolean
                        example: true
                      conflictedData:
                        type: array
                        items:
                          type: object
                          properties:
                            field:
                              type: string
                              example: "subcategory"
                            input:
                              type: object
                              properties:
                                ko:
                                  type: string
                                  example: "서브카테고리명(한)"
                                en:
                                  type: string
                                  example: "subcategory_name_en"
                        conflict:
                          type: array
                          items:
                            type: object
                            properties:
                              ko:
                                type: string
                                example: "conflict_value_ko"
                              en:
                                type: string
                                example: "conflict_value_en"
        '500':
          description: 서버 에러
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    example: 500
                  message:
                    type: string
                    example: "Internal server error"
...
  /api/seniorfaqs/update/category/conflict:
    post:
      summary: Senior FAQ 수정시 카테고리 충돌 여부 확인
      description: 수정하려는 Senior FAQ의 카테고리 충돌 여부를 확인하는 API입니다.
      operationId: updateCheckSeniorFAQCategoryConflict
      tags:
        - seniorfaqs
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                senior_faq_id:
                  type: integer
                  example: 1
                maincategory_ko:
                  type: string
                  example: "생활"
                maincategory_en:
                  type: string
                  example: "Life"
                subcategory_ko:
                  type: string
                  example: "건강"
                subcategory_en:
                  type: string
                  example: "Health"
                detailcategory_ko:
                  type: string
                  example: "노화 방지"
                detailcategory_en:
                  type: string
                  example: "Anti-aging"
      responses:
        '200':
          description: 카테고리 확인 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: "Categories checked successfully"
                  data:
                    type: object
                    properties:
                      isConflict:
                        type: boolean
                        example: true
                      conflictedData:
                        type: array
                        items:
                          type: object
                          properties:
                            field:
                              type: string
                              example: "subcategory"
                            input:
                              type: object
                              properties:
                                ko:
                                  type: string
                                  example: "서브카테고리명(한)"
                                en:
                                  type: string
                                  example: "subcategory_name_en"
                            conflict:
                              type: array
                              items:
                                type: object
                                properties:
                                  ko:
                                    type: string
                                    example: "conflict_value_ko"
                                  en:
                                    type: string
                                    example: "conflict_value_en"
        '500':
          description: 서버 오류 발생
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    example: 500
                  message:
                    type: string
                    example: "Internal server error"
...